Feat: add getAwsIdFromArn
Simple util to extract the last section of an ARN, which is typically
(but sadly not always) the ID of the resource.

For services, the result will be the name of the resource (S3 Bucket,
Function, etc.)

For executions (Step Functions, Lambda, etc.)  it'll be the ID of the
execution.